+++ examples/windows-recipe-writefile-advanced.rs
@@ -55,7 +55,7 @@ use vmi::{
     arch::amd64::Amd64,
     driver::VmiFullDriver,
     os::{VmiOsProcess as _, windows::WindowsOs},
-    utils::injector::{InjectorHandler, Recipe, RecipeControlFlow, recipe},
+    utils::injector::{Recipe, RecipeControlFlow, UserInjectorHandler, recipe},
 };
 
 #[derive(Debug, Default)]
@@ -330,14 +330,14 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     }
 
     session.handle(|session| {
-        InjectorHandler::new(
+        UserInjectorHandler::new(
             session,
-            explorer_pid,
             recipe_factory(GuestFile::new(
                 "C:\\Users\\John\\Desktop\\test.txt",
                 content,
             )),
-        )
+        )?
+        .with_pid(explorer_pid)
     })?;
 
     Ok(())